Ensure consistency in default chat channel
There was an inconsistency where sometimes the default channel would be
NULL and sometimes an empty string. This was not only causing NPEs, but
I think this could be the reason chat channels stopped working.
Hopefully I've caught all the places it matters. Needs testing.

This change doesn't need an update on proxy, only Bukkit servers.